+++ login-workflow/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   A `BrandedCardContainer` component.
 -   Confirm password field error state message in `ChangePasswordForm`.
+-   A `ExistingAccountRegisterSuccess` component.
 
 ### Changed